fix: make workflow host API first-class
