+++ docs/public/md/extend/workflows-v2.md
@@ -69,25 +69,23 @@ Supported v2 primitives:
 
 ### Keep imports narrow
 
-Workflow files should import only the workflow context compatibility module:
+Workflow files should import only the supported workflow-host API surface they
+need:
 
 ```python
 from api.workflow_engine import WorkflowContext
+from api.runtime_control import ControlPlaneError
 ```
 
-Do not import Python API internals such as:
+Supported workflow-host modules are `api.workflow_engine`,
+`api.runtime_control`, `api.app`, and `api.metrics`.
 
-```python
-from api.runtime_control import canonical_json
-from api.vm_metrics import workflow_counter
-```
-
-Those modules were implementation details of the Python API service. In v2,
-the workflow host provides a small compatibility surface instead of the whole
-Python API package.
+Do not import unrelated API-service internals or another workflow domain's local
+helpers. Domain-specific helpers should live next to the workflows that own
+them, for example `workflows/slack/metrics.py`.
 
 If a workflow needs a helper, move it into the workflow file, a shared overlay
-module, or a supported workflow-host compatibility shim.
+module, or a supported workflow-host API module.
 
 ### Put side effects behind steps
 
@@ -194,10 +192,10 @@ For each existing workflow:
 
 ## Known gaps
 
-The v2 POC supports the workflow model, but it does not yet emulate the full
-Python API package. Workflows that import `api.runtime_control`, `api.vm_metrics`,
-or other Python API internals need a compatibility shim or a small local helper
-before they are v2-ready.
+The v2 workflow host intentionally exposes a narrow Python API package.
+Workflows that import unrelated API-service internals should move that behavior
+into the workflow-host API surface or a small local helper owned by the workflow
+domain before they are v2-ready.
 
 The tool runtime is also still proxied. `ctx.call_tool(...)` works through the
 configured tool API, but a fully native `api-rs` tool runtime is a separate
