+++ services/workflow-python/api/__init__.py
@@ -1,0 +1,5 @@
+"""First-class Python workflow runtime API.
+
+This package is intentionally small: it contains the stable surface that
+workflow files can import when they run under the api-rs workflow host.
+"""
