+++ services/workflow-python/pyproject.toml
@@ -5,11 +5,13 @@ requires-python = ">=3.11"
 dependencies = [
     "asyncpg>=0.30.0",
     "boto3>=1.40.0",
+    "feedparser>=6.0.0",
     "google-api-python-client>=2.100.0",
     "google-auth-httplib2>=0.2.0",
     "google-auth-oauthlib>=1.2.0",
     "httplib2>=0.20.0",
     "httpx>=0.28.0",
     "pysocks>=1.7.1",
+    "rich>=13.0.0",
     "slack-sdk>=3.39.0",
 ]
