+++ services/workflow-python/workflow_host.py
@@ -15,98 +15,18 @@ import importlib.util
 import inspect
 import json
 import os
-import shutil
-import subprocess
 import sys
 import traceback
-import types
 import typing
-import urllib.request
 from pathlib import Path
 from typing import Any
+
+from api import metrics
+from api.workflow_engine import WorkflowContext
 
 
 class ProtocolError(RuntimeError):
     pass
-
-
-class WorkflowContext:
-    def __init__(
-        self,
-        rpc: "RpcClient",
-        *,
-        run_id: str,
-        task_id: str,
-        workflow_name: str,
-        pool: Any = None,
-    ) -> None:
-        self._rpc = rpc
-        self.run_id = run_id
-        self.task_id = task_id
-        self.workflow_name = workflow_name
-        self._pool = pool
-
-    def log(self, event: str, **fields: Any) -> None:
-        self._rpc.notify(
-            {
-                "type": "ctx.log",
-                "message": event,
-                "fields": fields,
-            }
-        )
-
-    async def step(self, name: str, fn: Any, *, retry: Any = None, timeout: Any = None) -> Any:
-        del retry, timeout
-        started = await self._rpc.request({"type": "ctx.step.get", "step": name})
-        if started.get("done"):
-            return started.get("value")
-
-        value = fn()
-        if inspect.isawaitable(value):
-            value = await value
-        await self._rpc.request(
-            {
-                "type": "ctx.step.put",
-                "checkpoint_name": started["checkpoint_name"],
-                "value": value,
-            }
-        )
-        return value
-
-    async def agent_turn(self, text: str | None = None, **kwargs: Any) -> Any:
-        args = dict(kwargs)
-        if text is not None:
-            args.setdefault("text", text)
-        return await self._rpc.request({"type": "ctx.agent_turn", "args": args})
-
-    async def run_agent(self, text: str | None = None, **kwargs: Any) -> Any:
-        return await self.agent_turn(text, **kwargs)
-
-    async def call_tool(self, tool: str, method: str, args: dict[str, Any] | None = None) -> Any:
-        tool_shim = resolve_tool_shim()
-        if tool_shim is not None:
-            # Sandboxed workflow hosts cannot rely on api-rs having a /tools
-            # backend. Use the generated catalog's method bridge for durable
-            # workflow ctx.call_tool(...); interactive agents use tool CLIs.
-            return await call_tool_shim(tool_shim, tool, method, args or {})
-        return await self._rpc.request(
-            {
-                "type": "ctx.call_tool",
-                "tool": tool,
-                "method": method,
-                "args": args or {},
-            }
-        )
-
-    async def post_to_slack(self, channel: str, text: str, **kwargs: Any) -> Any:
-        return await self._rpc.request(
-            {
-                "type": "ctx.post_to_slack",
-                "channel": channel,
-                "text": text,
-                "args": kwargs,
-            }
-        )
 
 
 class RpcClient:
@@ -152,53 +72,6 @@ class RpcClient:
             fut.set_exception(RuntimeError(str(response.get("error") or "context RPC failed")))
 
 
-_METRIC_RPC: RpcClient | None = None
-
-
-def resolve_tool_shim() -> str | None:
-    if tool_shim := shutil.which("centaur-tools"):
-        return tool_shim
-    fallback = Path("/home/agent/.local/bin/centaur-tools")
-    if fallback.exists():
-        return str(fallback)
-    installer = Path("/usr/local/bin/install-tool-shims")
-    if installer.exists():
-        subprocess.run(
-            [str(installer)],
-            check=False,
-            stdout=sys.stderr,
-            stderr=sys.stderr,
-        )
-        if tool_shim := shutil.which("centaur-tools"):
-            return tool_shim
-        if fallback.exists():
-            return str(fallback)
-    return None
-
-
-async def call_tool_shim(
-    tool_shim: str, tool: str, method: str, args: dict[str, Any]
-) -> Any:
-    proc = await asyncio.create_subprocess_exec(
-        tool_shim,
-        "call",
-        tool,
-        method,
-        json.dumps(args, separators=(",", ":"), default=str),
-        stdout=asyncio.subprocess.PIPE,
-        stderr=asyncio.subprocess.PIPE,
-    )
-    stdout, stderr = await proc.communicate()
-    text = stdout.decode(errors="replace").strip()
-    err = stderr.decode(errors="replace").strip()
-    if proc.returncode != 0:
-        detail = err or text or f"exit code {proc.returncode}"
-        raise RuntimeError(f"centaur-tools call {tool}.{method} failed: {detail}")
-    if not text:
-        return None
-    return json.loads(text)
-
-
 @dataclasses.dataclass
 class RegisteredWorkflow:
     workflow_name: str
@@ -207,860 +80,6 @@ class RegisteredWorkflow:
     input_cls: type | None
     webhooks: Any
     schedule: Any
-
-
-def install_api_compat_module() -> None:
-    api_mod = sys.modules.get("api")
-    if api_mod is None:
-        try:
-            import api as imported_api  # type: ignore
-
-            api_mod = imported_api
-        except ImportError:
-            api_mod = types.ModuleType("api")
-            api_mod.__path__ = []  # Mark as package so compat submodules can import.
-            sys.modules["api"] = api_mod
-
-    workflow_engine = types.ModuleType("api.workflow_engine")
-    workflow_engine.WorkflowContext = WorkflowContext
-    sys.modules["api.workflow_engine"] = workflow_engine
-    setattr(api_mod, "workflow_engine", workflow_engine)
-
-    runtime_control = types.ModuleType("api.runtime_control")
-    runtime_control.canonical_json = canonical_json
-    runtime_control.decode_jsonb = decode_jsonb
-    sys.modules.setdefault("api.runtime_control", runtime_control)
-    setattr(api_mod, "runtime_control", runtime_control)
-
-    install_vm_metrics_compat_module(api_mod)
-
-    install_centaur_sdk_compat_module()
-
-
-def canonical_json(value: Any) -> str:
-    return json.dumps(value, sort_keys=True, separators=(",", ":"), default=str)
-
-
-def decode_jsonb(value: Any, fallback: Any) -> Any:
-    if value is None:
-        return fallback
-    if isinstance(value, str):
-        try:
-            return json.loads(value)
-        except json.JSONDecodeError:
-            return fallback
-    return value
-
-
-def install_centaur_sdk_compat_module() -> None:
-    if "centaur_sdk" in sys.modules:
-        return
-    try:
-        __import__("centaur_sdk")
-        return
-    except ImportError:
-        pass
-    centaur_sdk = types.ModuleType("centaur_sdk")
-
-    def secret(name: str, default: str | None = None) -> str:
-        return os.getenv(name, default or "")
-
-    centaur_sdk.secret = secret
-    sys.modules["centaur_sdk"] = centaur_sdk
-
-
-_METRIC_COUNTERS: dict[tuple[str, tuple[tuple[str, str], ...]], float] = {}
-_METRIC_GAUGES: dict[tuple[str, tuple[tuple[str, str], ...]], float] = {}
-_METRIC_HISTOGRAMS: dict[tuple[str, tuple[tuple[str, str], ...]], dict[str, Any]] = {}
-_COMPANY_CONTEXT_DOCUMENT_SIZE_BUCKETS = [
-    100,
-    500,
-    1_000,
-    5_000,
-    10_000,
-    25_000,
-    50_000,
-    100_000,
-    250_000,
-    500_000,
-]
-_SLACK_ARCHIVE_IMPORT_BATCH_SIZE_BUCKETS = [
-    1,
-    10,
-    100,
-    500,
-    1_000,
-    5_000,
-    10_000,
-]
-_SLACK_ARCHIVE_IMPORT_DURATION_BUCKETS = [
-    1,
-    5,
-    10,
-    30,
-    60,
-    120,
-    300,
-    600,
-    1_200,
-    3_600,
-]
-_SLACK_RETENTION_DURATION_BUCKETS = [
-    1,
-    5,
-    10,
-    30,
-    60,
-    120,
-    300,
-    600,
-    1_200,
-]
-
-
-def install_vm_metrics_compat_module(api_mod: types.ModuleType) -> None:
-    if "api.vm_metrics" in sys.modules:
-        setattr(api_mod, "vm_metrics", sys.modules["api.vm_metrics"])
-        return
-    try:
-        import api.vm_metrics as vm_metrics  # type: ignore
-
-        setattr(api_mod, "vm_metrics", vm_metrics)
-        return
-    except ImportError:
-        pass
-
-    vm_metrics = types.ModuleType("api.vm_metrics")
-    vm_metrics.record_etl_items_deleted = record_etl_items_deleted
-    vm_metrics.record_etl_items_enqueued = record_etl_items_enqueued
-    vm_metrics.record_etl_items_failed = record_etl_items_failed
-    vm_metrics.record_etl_items_seen = record_etl_items_seen
-    vm_metrics.record_etl_items_upserted = record_etl_items_upserted
-    vm_metrics.record_slack_etl_rate_limit = record_slack_etl_rate_limit
-    vm_metrics.record_slack_archive_import_batch_failure = (
-        record_slack_archive_import_batch_failure
-    )
-    vm_metrics.record_slack_archive_import_batch_size = (
-        record_slack_archive_import_batch_size
-    )
-    vm_metrics.record_slack_archive_import_bytes = record_slack_archive_import_bytes
-    vm_metrics.record_slack_archive_import_channels = (
-        record_slack_archive_import_channels
-    )
-    vm_metrics.record_slack_archive_import_failure = (
-        record_slack_archive_import_failure
-    )
-    vm_metrics.record_slack_archive_import_message_files = (
-        record_slack_archive_import_message_files
-    )
-    vm_metrics.record_slack_archive_import_messages = (
-        record_slack_archive_import_messages
-    )
-    vm_metrics.record_slack_archive_import_attachments = (
-        record_slack_archive_import_attachments
-    )
-    vm_metrics.record_slack_archive_import_run = record_slack_archive_import_run
-    vm_metrics.record_slack_archive_import_skipped_items = (
-        record_slack_archive_import_skipped_items
-    )
-    vm_metrics.record_slack_archive_import_users = record_slack_archive_import_users
-    vm_metrics.observe_slack_archive_import_batch_duration = (
-        observe_slack_archive_import_batch_duration
-    )
-    vm_metrics.observe_slack_archive_import_duration = (
-        observe_slack_archive_import_duration
-    )
-    vm_metrics.record_slack_retention_backfill_job = (
-        record_slack_retention_backfill_job
-    )
-    vm_metrics.record_slack_retention_backfill_job_failure = (
-        record_slack_retention_backfill_job_failure
-    )
-    vm_metrics.record_slack_retention_backfill_terminal_skip = (
-        record_slack_retention_backfill_terminal_skip
-    )
-    vm_metrics.record_slack_retention_channel_failure = (
-        record_slack_retention_channel_failure
-    )
-    vm_metrics.record_slack_retention_failure = record_slack_retention_failure
-    vm_metrics.record_slack_retention_messages_processed = (
-        record_slack_retention_messages_processed
-    )
-    vm_metrics.record_slack_retention_api_request = (
-        record_slack_retention_api_request
-    )
-    vm_metrics.record_slack_retention_api_rate_limited = (
-        record_slack_retention_api_rate_limited
-    )
-    vm_metrics.record_slack_retention_run = record_slack_retention_run
-    vm_metrics.observe_slack_retention_run_duration = (
-        observe_slack_retention_run_duration
-    )
-    vm_metrics.set_slack_archive_import_last_failure_timestamp = (
-        set_slack_archive_import_last_failure_timestamp
-    )
-    vm_metrics.set_slack_retention_last_failure_timestamp = (
-        set_slack_retention_last_failure_timestamp
-    )
-    vm_metrics.set_slack_retention_watermark_lag_seconds = (
-        set_slack_retention_watermark_lag_seconds
-    )
-    vm_metrics.set_etl_active_scopes = set_etl_active_scopes
-    vm_metrics.set_etl_backfill_job_age_seconds = set_etl_backfill_job_age_seconds
-    vm_metrics.set_etl_backfill_jobs = set_etl_backfill_jobs
-    vm_metrics.set_etl_failed_scopes = set_etl_failed_scopes
-    vm_metrics.set_etl_scope_sync_freshness_seconds = (
-        set_etl_scope_sync_freshness_seconds
-    )
-    vm_metrics.record_company_context_documents_changed = (
-        record_company_context_documents_changed
-    )
-    vm_metrics.observe_company_context_document_size = (
-        observe_company_context_document_size
-    )
-    vm_metrics.set_company_context_projection_lag = (
-        set_company_context_projection_lag
-    )
-    sys.modules["api.vm_metrics"] = vm_metrics
-    setattr(api_mod, "vm_metrics", vm_metrics)
-
-
-def record_etl_items_seen(
-    source: str, source_type: str, item_type: str, count: int
-) -> None:
-    increment_metric(
-        "etl_items_seen_total",
-        count,
-        source=source,
-        source_type=source_type,
-        item_type=item_type,
-    )
-
-
-def record_etl_items_enqueued(
-    source: str, source_type: str, item_type: str, count: int
-) -> None:
-    increment_metric(
-        "etl_items_enqueued_total",
-        count,
-        source=source,
-        source_type=source_type,
-        item_type=item_type,
-    )
-
-
-def record_etl_items_upserted(
-    source: str, source_type: str, item_type: str, count: int
-) -> None:
-    increment_metric(
-        "etl_items_upserted_total",
-        count,
-        source=source,
-        source_type=source_type,
-        item_type=item_type,
-    )
-
-
-def record_etl_items_deleted(
-    source: str, source_type: str, item_type: str, count: int
-) -> None:
-    increment_metric(
-        "etl_items_deleted_total",
-        count,
-        source=source,
-        source_type=source_type,
-        item_type=item_type,
-    )
-
-
-def record_etl_items_failed(
-    source: str,
-    source_type: str,
-    item_type: str,
-    reason: str,
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "etl_items_failed_total",
-        count,
-        source=source,
-        source_type=source_type,
-        item_type=item_type,
-        reason=reason,
-    )
-
-
-def record_slack_etl_rate_limit(
-    workflow: str,
-    method: str,
-    outcome: str,
-    retry_after_seconds: int | float,
-) -> None:
-    retry_after = max(float(retry_after_seconds), 0.0)
-    labels = {
-        "workflow": workflow,
-        "method": method,
-        "outcome": outcome,
-    }
-    increment_metric("slack_etl_rate_limits_total", 1, **labels)
-    increment_metric(
-        "slack_etl_rate_limit_retry_after_seconds_total",
-        retry_after,
-        **labels,
-    )
-
-
-def record_slack_archive_import_run(
-    status: str,
-    reason: str = "none",
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "slack_archive_import_runs_total",
-        count,
-        status=status,
-        reason=reason,
-    )
-
-
-def observe_slack_archive_import_duration(status: str, duration_s: float) -> None:
-    observe_histogram(
-        "slack_archive_import_duration_seconds",
-        max(duration_s, 0.0),
-        _SLACK_ARCHIVE_IMPORT_DURATION_BUCKETS,
-        status=status,
-    )
-
-
-def record_slack_archive_import_bytes(count: int) -> None:
-    increment_metric("slack_archive_import_bytes_total", count)
-
-
-def record_slack_archive_import_channels(result: str, count: int) -> None:
-    increment_metric("slack_archive_import_channels_total", count, result=result)
-
-
-def record_slack_archive_import_users(result: str, count: int) -> None:
-    increment_metric("slack_archive_import_users_total", count, result=result)
-
-
-def record_slack_archive_import_messages(result: str, count: int) -> None:
-    increment_metric("slack_archive_import_messages_total", count, result=result)
-
-
-def record_slack_archive_import_message_files(result: str, count: int) -> None:
-    increment_metric("slack_archive_import_message_files_total", count, result=result)
-
-
-def record_slack_archive_import_attachments(result: str, count: int) -> None:
-    increment_metric("slack_archive_import_attachments_total", count, result=result)
-
-
-def observe_slack_archive_import_batch_duration(entity: str, duration_s: float) -> None:
-    observe_histogram(
-        "slack_archive_import_batch_duration_seconds",
-        max(duration_s, 0.0),
-        _SLACK_ARCHIVE_IMPORT_DURATION_BUCKETS,
-        entity=entity,
-    )
-
-
-def record_slack_archive_import_batch_size(entity: str, count: int) -> None:
-    observe_histogram(
-        "slack_archive_import_batch_size",
-        max(count, 0),
-        _SLACK_ARCHIVE_IMPORT_BATCH_SIZE_BUCKETS,
-        entity=entity,
-    )
-
-
-def record_slack_archive_import_failure(
-    stage: str,
-    reason: str,
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "slack_archive_import_failures_total",
-        count,
-        stage=stage,
-        reason=reason,
-    )
-
-
-def record_slack_archive_import_skipped_items(
-    item_type: str,
-    reason: str,
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "slack_archive_import_skipped_items_total",
-        count,
-        item_type=item_type,
-        reason=reason,
-    )
-
-
-def record_slack_archive_import_batch_failure(
-    entity: str,
-    reason: str,
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "slack_archive_import_batch_failures_total",
-        count,
-        entity=entity,
-        reason=reason,
-    )
-
-
-def set_slack_archive_import_last_failure_timestamp(timestamp_s: float) -> None:
-    set_gauge("slack_archive_import_last_failure_timestamp_seconds", timestamp_s)
-
-
-def record_slack_retention_run(
-    workflow: str,
-    status: str,
-    mode: str,
-    reason: str = "none",
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "slack_retention_runs_total",
-        count,
-        workflow=workflow,
-        status=status,
-        mode=mode,
-        reason=reason,
-    )
-
-
-def observe_slack_retention_run_duration(
-    workflow: str,
-    mode: str,
-    status: str,
-    duration_s: float,
-) -> None:
-    observe_histogram(
-        "slack_retention_run_duration_seconds",
-        max(duration_s, 0.0),
-        _SLACK_RETENTION_DURATION_BUCKETS,
-        workflow=workflow,
-        mode=mode,
-        status=status,
-    )
-
-
-def record_slack_retention_messages_processed(
-    workflow: str,
-    mode: str,
-    result: str,
-    count: int,
-) -> None:
-    increment_metric(
-        "slack_retention_messages_processed_total",
-        count,
-        workflow=workflow,
-        mode=mode,
-        result=result,
-    )
-
-
-def record_slack_retention_backfill_job(
-    job_type: str,
-    result: str,
-    reason: str = "none",
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "slack_retention_backfill_jobs_total",
-        count,
-        job_type=job_type,
-        result=result,
-        reason=reason,
-    )
-
-
-def record_slack_retention_failure(
-    workflow: str,
-    operation: str,
-    reason: str,
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "slack_retention_failures_total",
-        count,
-        workflow=workflow,
-        operation=operation,
-        reason=reason,
-    )
-
-
-def record_slack_retention_api_request(
-    operation: str,
-    result: str,
-    reason: str = "none",
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "slack_retention_api_requests_total",
-        count,
-        operation=operation,
-        result=result,
-        reason=reason,
-    )
-
-
-def record_slack_retention_api_rate_limited(operation: str, count: int = 1) -> None:
-    increment_metric(
-        "slack_retention_api_rate_limited_total",
-        count,
-        operation=operation,
-    )
-
-
-def record_slack_retention_backfill_job_failure(
-    job_type: str,
-    reason: str,
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "slack_retention_backfill_job_failures_total",
-        count,
-        job_type=job_type,
-        reason=reason,
-    )
-
-
-def record_slack_retention_backfill_terminal_skip(
-    job_type: str,
-    reason: str,
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "slack_retention_backfill_terminal_skips_total",
-        count,
-        job_type=job_type,
-        reason=reason,
-    )
-
-
-def record_slack_retention_channel_failure(
-    workflow: str,
-    reason: str,
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "slack_retention_channel_failures_total",
-        count,
-        workflow=workflow,
-        reason=reason,
-    )
-
-
-def set_slack_retention_last_failure_timestamp(workflow: str, timestamp_s: float) -> None:
-    set_gauge(
-        "slack_retention_last_failure_timestamp_seconds",
-        timestamp_s,
-        workflow=workflow,
-    )
-
-
-def set_slack_retention_watermark_lag_seconds(mode: str, lag_s: float) -> None:
-    set_gauge(
-        "slack_retention_watermark_lag_seconds",
-        max(lag_s, 0.0),
-        mode=mode,
-    )
-
-
-def set_etl_active_scopes(source: str, count: int) -> None:
-    set_gauge(
-        "etl_active_scopes",
-        max(count, 0),
-        source=source,
-    )
-
-
-def set_etl_failed_scopes(source: str, count: int) -> None:
-    set_gauge(
-        "etl_failed_scopes",
-        max(count, 0),
-        source=source,
-    )
-
-
-def set_etl_scope_sync_freshness_seconds(
-    source: str,
-    freshness_s: int | float,
-) -> None:
-    set_gauge(
-        "etl_scope_sync_freshness_seconds",
-        max(float(freshness_s), 0.0),
-        source=source,
-    )
-
-
-def set_etl_backfill_jobs(
-    source: str,
-    job_type: str,
-    status: str,
-    count: int,
-) -> None:
-    set_gauge(
-        "etl_backfill_jobs",
-        max(count, 0),
-        source=source,
-        job_type=job_type,
-        status=status,
-    )
-
-
-def set_etl_backfill_job_age_seconds(
-    source: str,
-    job_type: str,
-    status: str,
-    age_s: int | float,
-) -> None:
-    set_gauge(
-        "etl_backfill_job_age_seconds",
-        max(float(age_s), 0.0),
-        source=source,
-        job_type=job_type,
-        status=status,
-    )
-
-
-def record_company_context_documents_changed(
-    source: str,
-    source_type: str,
-    action: str,
-    count: int = 1,
-) -> None:
-    increment_metric(
-        "company_context_documents_changed_total",
-        count,
-        source=source,
-        source_type=source_type,
-        action=action,
-    )
-
-
-def observe_company_context_document_size(
-    source: str, source_type: str, chars: int
-) -> None:
-    observe_histogram(
-        "company_context_document_size_chars",
-        max(chars, 0),
-        _COMPANY_CONTEXT_DOCUMENT_SIZE_BUCKETS,
-        source=source,
-        source_type=source_type,
-    )
-
-
-def set_company_context_projection_lag(source: str, projection_lag_s: float) -> None:
-    set_gauge(
-        "company_context_projection_lag_seconds",
-        max(projection_lag_s, 0.0),
-        source=source,
-    )
-
-
-def increment_metric(metric: str, count: int, **labels: str) -> None:
-    if count < 0:
-        return
-    labels = metric_runtime_labels(labels)
-    key = metric_key(metric, labels)
-    _METRIC_COUNTERS[key] = _METRIC_COUNTERS.get(key, 0.0) + float(count)
-    emit_metric_event("counter", metric, count, labels)
-    push_metric_lines([format_metric_line(metric, labels, _METRIC_COUNTERS[key])])
-
-
-def set_gauge(metric: str, value: float, **labels: str) -> None:
-    labels = metric_runtime_labels(labels)
-    key = metric_key(metric, labels)
-    _METRIC_GAUGES[key] = float(value)
-    emit_metric_event("gauge", metric, float(value), labels)
-    push_metric_lines([format_metric_line(metric, labels, _METRIC_GAUGES[key])])
-
-
-def observe_histogram(
-    metric: str,
-    value: int | float,
-    buckets: list[int],
-    **labels: str,
-) -> None:
-    labels = metric_runtime_labels(labels)
-    key = metric_key(metric, labels)
-    histogram = _METRIC_HISTOGRAMS.setdefault(
-        key,
-        {
-            "buckets": {bucket: 0 for bucket in buckets},
-            "count": 0,
-            "sum": 0.0,
-        },
-    )
-    numeric = float(value)
-    histogram["count"] += 1
-    histogram["sum"] += numeric
-    emit_metric_event("histogram", metric, numeric, labels)
-    for bucket in buckets:
-        if numeric <= bucket:
-            histogram["buckets"][bucket] += 1
-
-    lines = []
-    for bucket in buckets:
-        lines.append(
-            format_metric_line(
-                f"{metric}_bucket",
-                {**labels, "le": str(float(bucket))},
-                histogram["buckets"][bucket],
-            )
-        )
-    lines.append(
-        format_metric_line(
-            f"{metric}_bucket",
-            {**labels, "le": "+Inf"},
-            histogram["count"],
-        )
-    )
-    lines.append(format_metric_line(f"{metric}_count", labels, histogram["count"]))
-    lines.append(format_metric_line(f"{metric}_sum", labels, histogram["sum"]))
-    push_metric_lines(lines)
-
-
-def metric_key(
-    metric: str, labels: dict[str, str]
-) -> tuple[str, tuple[tuple[str, str], ...]]:
-    return (metric, tuple(sorted((key, str(value)) for key, value in labels.items())))
-
-
-def emit_metric_event(
-    kind: str, metric: str, value: int | float, labels: dict[str, str]
-) -> None:
-    if _METRIC_RPC is None:
-        return
-    _METRIC_RPC.notify(
-        {
-            "type": "ctx.metric",
-            "kind": kind,
-            "name": metric,
-            "value": value,
-            "labels": labels,
-        }
-    )
-
-
-def metric_runtime_labels(labels: dict[str, str]) -> dict[str, str]:
-    runtime_labels = dict(labels)
-    for key, value in default_metric_runtime_labels().items():
-        runtime_labels.setdefault(key, value)
-    return runtime_labels
-
-
-def default_metric_runtime_labels() -> dict[str, str]:
-    labels: dict[str, str] = {}
-    namespace = runtime_namespace()
-    if namespace:
-        labels["namespace"] = namespace
-    environment = runtime_environment()
-    if environment:
-        labels["environment"] = environment
-    return labels
-
-
-def runtime_namespace() -> str | None:
-    for name in (
-        "METRICS_NAMESPACE",
-        "KUBERNETES_NAMESPACE",
-        "POD_NAMESPACE",
-        "SESSION_SANDBOX_K8S_NAMESPACE",
-    ):
-        value = clean_metric_label_value(os.environ.get(name))
-        if value:
-            return value
-
-    try:
-        namespace = Path("/var/run/secrets/kubernetes.io/serviceaccount/namespace")
-        if namespace.exists():
-            return clean_metric_label_value(namespace.read_text(encoding="utf-8"))
-    except OSError:
-        return None
-    return None
-
-
-def runtime_environment() -> str | None:
-    for name in ("METRICS_ENVIRONMENT", "ENVIRONMENT", "DEPLOYMENT_ENVIRONMENT"):
-        value = clean_metric_label_value(os.environ.get(name))
-        if value:
-            return value
-
-    for attr in os.environ.get("OTEL_RESOURCE_ATTRIBUTES", "").split(","):
-        key, separator, value = attr.partition("=")
-        if separator and key.strip() in {
-            "deployment.environment",
-            "deployment.environment.name",
-        }:
-            cleaned = clean_metric_label_value(value)
-            if cleaned:
-                return cleaned
-
-    namespace = runtime_namespace()
-    if namespace == "centaur-system":
-        return "production"
-    if namespace and namespace.startswith("stg-"):
-        return "staging"
-    return None
-
-
-def clean_metric_label_value(value: str | None) -> str | None:
-    if value is None:
-        return None
-    value = value.strip()
-    return value or None
-
-
-def format_metric_line(metric: str, labels: dict[str, str], value: float) -> str:
-    if labels:
-        label_text = ",".join(
-            f'{key}="{escape_label_value(str(label_value))}"'
-            for key, label_value in sorted(labels.items())
-        )
-        return f"{metric}{{{label_text}}} {value}"
-    return f"{metric} {value}"
-
-
-def escape_label_value(value: str) -> str:
-    return value.replace("\\", "\\\\").replace('"', '\\"').replace("\n", "\\n")
-
-
-def push_metric_lines(lines: list[str]) -> None:
-    if not victoria_metrics_push_enabled():
-        return
-    payload = ("\n".join(lines) + "\n").encode("utf-8")
-    request = urllib.request.Request(
-        f"{victoria_metrics_url().rstrip('/')}/api/v1/import/prometheus",
-        data=payload,
-        headers={"Content-Type": "text/plain"},
-        method="POST",
-    )
-    try:
-        opener = urllib.request.build_opener(urllib.request.ProxyHandler({}))
-        opener.open(request, timeout=2).close()
-    except Exception as exc:
-        print(f"workflow_metric_push_error error={exc}", file=sys.stderr)
-
-
-def victoria_metrics_url() -> str:
-    return os.environ.get("VICTORIAMETRICS_URL", "http://victoriametrics:8428")
-
-
-def victoria_metrics_push_enabled() -> bool:
-    return os.environ.get("VICTORIAMETRICS_PUSH_ENABLED", "1").strip().lower() not in {
-        "0",
-        "false",
-        "no",
-        "off",
-    }
 
 
 def workflow_dirs() -> list[Path]:
@@ -1149,7 +168,6 @@ def has_workflow_name_assignment(path: Path) -> bool:
 def discover_workflows() -> dict[str, RegisteredWorkflow]:
     dirs = workflow_dirs()
     configure_workflow_import_paths(dirs)
-    install_api_compat_module()
     discovered: dict[str, RegisteredWorkflow] = {}
     for directory in dirs:
         for path in sorted(directory.rglob("*.py")):
@@ -1278,7 +296,6 @@ def normalize_schedule(workflow: RegisteredWorkflow) -> dict[str, Any] | None:
 
 
 async def run_workflow(message: dict[str, Any], rpc: RpcClient) -> dict[str, Any]:
-    global _METRIC_RPC
     workflows = discover_workflows()
     workflow_name = str(message.get("workflow_name") or "")
     registered = workflows.get(workflow_name)
@@ -1293,8 +310,8 @@ async def run_workflow(message: dict[str, Any], rpc: RpcClient) -> dict[str, Any
         workflow_name=workflow_name,
         pool=pool,
     )
-    previous_metric_rpc = _METRIC_RPC
-    _METRIC_RPC = rpc
+    previous_metric_rpc = metrics.get_metric_rpc()
+    metrics.set_metric_rpc(rpc)
     try:
         inp = coerce_input(message.get("input") or {}, registered.input_cls)
         result = registered.handler(inp, ctx)
@@ -1311,7 +328,7 @@ async def run_workflow(message: dict[str, Any], rpc: RpcClient) -> dict[str, Any
         }
     finally:
         await rpc.drain_notifications()
-        _METRIC_RPC = previous_metric_rpc
+        metrics.set_metric_rpc(previous_metric_rpc)
         if pool is not None:
             await pool.close()
