+++ workflows/linear/sync.py
@@ -9,7 +9,7 @@ from dataclasses import dataclass, field
 from typing import Any, Protocol
 
 from api.runtime_control import canonical_json
-from api.vm_metrics import (
+from workflows.etl_metrics import (
     record_etl_items_failed,
     record_etl_items_seen,
     record_etl_items_upserted,
