+++ workflows/tests/test_company_context_documents_attachments.py
@@ -23,26 +23,31 @@ def _load_projection_module():
         value if value is not None else default
     )
 
-    vm_metrics = types.ModuleType("api.vm_metrics")
+    company_context_metrics = types.ModuleType("workflows.company_context_metrics")
     for name in (
         "observe_company_context_document_size",
         "record_company_context_documents_changed",
         "set_company_context_projection_lag",
+    ):
+        setattr(company_context_metrics, name, lambda *_args, **_kwargs: None)
+
+    etl_metrics = types.ModuleType("workflows.etl_metrics")
+    for name in (
         "set_etl_active_scopes",
         "set_etl_failed_scopes",
         "set_etl_scope_sync_freshness_seconds",
     ):
-        setattr(vm_metrics, name, lambda *_args, **_kwargs: None)
+        setattr(etl_metrics, name, lambda *_args, **_kwargs: None)
 
     workflow_engine = types.ModuleType("api.workflow_engine")
     workflow_engine.WorkflowContext = object
 
     api_module.runtime_control = runtime_control
-    api_module.vm_metrics = vm_metrics
     api_module.workflow_engine = workflow_engine
     sys.modules.setdefault("api", api_module)
     sys.modules.setdefault("api.runtime_control", runtime_control)
-    sys.modules["api.vm_metrics"] = vm_metrics
+    sys.modules["workflows.company_context_metrics"] = company_context_metrics
+    sys.modules["workflows.etl_metrics"] = etl_metrics
     sys.modules.setdefault("api.workflow_engine", workflow_engine)
 
     return importlib.import_module("workflows.company_context_documents")
